+++ packages/moonstone/Icon/Icon.js
@@ -46,8 +46,7 @@ const IconBase = kind({
 // Let's find a way to import this list directly, and bonus feature, render our icons in the docs
 // next to their names.
 /**
- * {@link moonstone/Icon.iconList} is an object whose keys can be used as the child of an
- * {@link moonstone/Icon.Icon} component.
+ * An object whose keys can be used as the child of an [Icon]{@link moonstone/Icon.Icon} component.
  *
  * List of Icons:
  * ```
@@ -136,7 +135,7 @@ const IconBase = kind({
  */
 
 /**
- * Moonstone-specific behaviors to apply to [Icon]{@link moonstone/Icon.IconBase}.
+ * Moonstone-specific behaviors to apply to [IconBase]{@link moonstone/Icon.IconBase}.
  *
  * @hoc
  * @memberof moonstone/Icon