+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Deprecated
+
+### Added
+
+### Changed
+
+### Fixed
+
+- `spotlight/Spottable` to prevent unnecessary updates due to focus changes
+
 ## [2.0.0-alpha.3] - 2018-01-18
 
 No significant changes.