+++ packages/spotlight/CHANGELOG.md
@@ -2,11 +2,13 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+- `spotlight/Spottable` to not remove `tabindex` from unspottable components to allow blur events to propagate as expected when a component becomes disabled
+
 ## [2.0.0-alpha.3] - 2018-01-18
 
 No significant changes.
-
-- `spotlight/Spottable` to not remove `tabindex` from unspottable components to allow blur events to propagate as expected when a component becomes disabled
 
 ## [2.0.0-alpha.2] - 2017-08-29